fix: resolve 176 compilation errors and warnings
- Remove unused imports across multiple files
- Add #[allow(dead_code)] for unused structs and functions
- Fix deprecated pyclass with from_py_object attribute
- Rename to_str to as_str to fix clippy wrong_self_convention
- Fix thread_local to use const {} for clippy
- Fix uppercase acronyms (GELU -> Gelu, SGD)
- Fix unnecessary_cast in dropout.rs
- Fix wildcard_in_or_patterns in loss.rs
- Add necessary #[allow(...)] for clippy warnings
- Fix manual_is_multiple_of in callbacks.rs
- Fix needlessly_borrow in tensor.rs

+++ src/nn/dropout.rs
@@ -23,7 +23,7 @@ impl Module for Dropout {
                 .to_numpy()
                 .iter()
                 .map(|&v| {
-                    if rand::random::<f64>() < (1.0 - self.p) as f64 {
+                    if rand::random::<f64>() < (1.0 - self.p) {
                         v / (1.0 - self.p) as f32
                     } else {
                         0.0
